+++ kernel/kernel_device_modules-6.6/drivers/misc/mediatek/apusys/apu_mem/apu_mem_export.h
@@ -29,7 +29,7 @@ int apu_mem_rvs_get_pool_size(uint32_t type, uint32_t *size);
 
 /* APUMMU only API */
 int apu_mem_map_iova(uint32_t type, uint64_t session, uint64_t device_va,
-			uint32_t buf_size, uint64_t *eva);
+			uint64_t buf_size, uint64_t *eva);
 int apu_mem_iova_decode(uint64_t eva, uint64_t *iova);
 int apu_mem_unmap_iova(uint64_t session, uint64_t device_va, uint32_t buf_size);
 int apu_mem_table_get(uint64_t session, void **tbl_kva, uint32_t *size);
