+++ kernel/kernel_device_modules-6.6/drivers/misc/mediatek/apusys/apu_mem/apummu/common/apummu_export.c
@@ -369,7 +369,7 @@ err:
  *  for apummu, we also record translate info into session table
  */
 int apummu_iova2eva(uint32_t type, uint64_t session, uint64_t device_va,
-		uint32_t buf_size, uint64_t *eva)
+		uint64_t buf_size, uint64_t *eva)
 {
 	return addr_encode_and_write_stable(type, session, device_va, buf_size, eva);
 }
